refactor(bridge): extract bridge_llm_stream module (Phase 2 batch 3)
Extract call_llm_stream, is_valid_tool_name, tool_call_start_event, bridge_llm_cancel,
rate_limit_cooldown, and related constants into turn/bridge_llm_stream.rs (~628 lines, 6 tests).

Remove 5 duplicate tests from bridge_inprocess.rs. Fix Uuid import.

bridge_inprocess.rs: 5920 → 4344 lines (cumulative -1576 across 3 batches).
